[MODFIN-141] - Split configuration
